removed interpreter2 file
