+++ src/p2/interpreter.rs
@@ -1,50 +1,29 @@
-pub(crate) struct Interpreter<'a> {
-    lib: &'a dyn ftd::p2::Library,
-    pub bag: std::collections::BTreeMap<String, ftd::p2::Thing>,
-    pub p1: Vec<ftd::p1::Section>,
-    pub aliases: std::collections::BTreeMap<String, String>,
-    pub parsed_libs: Vec<String>,
+#[derive(Debug, Default)]
+pub struct InterpreterState {
+    pub(crate) bag: std::collections::BTreeMap<String, ftd::p2::Thing>,
+    pub(crate) document_stack: Vec<ParsedDocument>,
+    pub(crate) parsed_libs: Vec<String>,
 }
 
-impl<'a> Interpreter<'a> {
-    #[cfg(feature = "async")]
-    pub(crate) fn interpret(
-        &mut self,
-        name: &str,
-        s: &str,
-    ) -> ftd::p1::Result<Vec<ftd::Instruction>> {
-        futures::executor::block_on(self.async_interpret(name, s))
+impl InterpreterState {
+    fn new() -> InterpreterState {
+        InterpreterState {
+            bag: ftd::p2::interpreter::default_bag(),
+            ..Default::default()
+        }
     }
 
-    // #[observed(with_result, namespace = "ftd")]
-    #[cfg(feature = "async")]
-    pub(crate) async fn async_interpret(
-        &mut self,
-        name: &str,
-        s: &str,
-    ) -> ftd::p1::Result<Vec<ftd::Instruction>> {
-        let mut d_get = std::time::Duration::new(0, 0);
-        let mut d_processor = std::time::Duration::new(0, 0);
-        let v = self
-            .async_interpret_(name, s, true, &mut d_get, &mut d_processor)
-            .await?;
-        // observer::observe_string("time_get", elapsed(d_get).as_str());
-        // observer::observe_string("time_processor", elapsed(d_processor).as_str());
-        Ok(v)
-    }
-
-    #[cfg(not(feature = "async"))]
-    pub(crate) fn interpret(
-        &mut self,
-        name: &str,
-        s: &str,
-    ) -> ftd::p1::Result<Vec<ftd::Instruction>> {
-        let mut d_get = std::time::Duration::new(0, 0);
-        let mut d_processor = std::time::Duration::new(0, 0);
-        let v = self.interpret_(name, s, true, &mut d_get, &mut d_processor)?;
-        // observer::observe_string("time_get", elapsed(d_get).as_str());
-        // observer::observe_string("time_processor", elapsed(d_processor).as_str());
-        Ok(v)
+    fn tdoc<'a>(
+        &'a self,
+        local_variables: &'a mut std::collections::BTreeMap<String, ftd::p2::Thing>,
+    ) -> ftd::p2::TDoc<'a> {
+        let l = self.document_stack.len() - 1;
+        ftd::p2::TDoc {
+            name: &self.document_stack[l].name,
+            aliases: &self.document_stack[l].doc_aliases,
+            bag: &self.bag,
+            local_variables,
+        }
     }
 
     fn library_in_the_bag(&self, name: &str) -> bool {
@@ -57,102 +36,42 @@ impl<'a> Interpreter<'a> {
         }
     }
 
-    #[cfg(feature = "async")]
-    #[async_recursion::async_recursion(?Send)]
-    async fn async_interpret_(
-        &mut self,
-        name: &str,
-        s: &str,
-        is_main: bool,
-        d_get: &mut std::time::Duration,
-        d_processor: &mut std::time::Duration,
-    ) -> ftd::p1::Result<Vec<ftd::Instruction>> {
-        let p1 = ftd::p1::parse(s, name)?;
-
-        let mut aliases = default_aliases();
-        let mut iteration_index = 0;
-        while iteration_index < p1.len() && p1[iteration_index].name == "import" {
-            if p1[iteration_index].is_commented {
-                iteration_index += 1;
-                continue;
-            }
-            let (library_name, alias) = ftd::p2::utils::parse_import(
-                &p1[iteration_index].caption,
-                name,
-                p1[iteration_index].line_number,
-            )?;
-            aliases.insert(alias, library_name.clone());
-            let start = std::time::Instant::now();
-            let doc = ftd::p2::TDoc {
-                name,
-                aliases: &aliases,
-                bag: &self.bag,
-                local_variables: &mut Default::default(),
-            };
-            let s = self
-                .lib
-                .get_with_result(library_name.as_str(), &doc)
-                .await?;
-            *d_get = d_get.saturating_add(std::time::Instant::now() - start);
-            if !self.library_in_the_bag(library_name.as_str()) {
-                self.async_interpret_(library_name.as_str(), s.as_str(), false, d_get, d_processor)
-                    .await?;
-                self.add_library_to_bag(library_name.as_str())
-            }
-            iteration_index += 1;
+    fn continue_(mut self) -> ftd::p1::Result<Interpreter> {
+        if self.document_stack.is_empty() {
+            panic!()
         }
-        let (new_p1, var_types) = ftd::p2::utils::reorder(
-            &p1[iteration_index..],
-            &ftd::p2::TDoc {
-                name,
-                aliases: &aliases,
-                bag: &self.bag,
-                local_variables: &mut Default::default(),
-            },
-        )?;
+
+        let l = self.document_stack.len() - 1; // Get the top of the stack
+
+        if self.document_stack[l].processing_imports {
+            let top = &mut self.document_stack[l];
+            let module = Self::process_imports(top, &self.bag)?;
+            if let Some(module) = module {
+                if !self.library_in_the_bag(module.as_str()) {
+                    self.add_library_to_bag(module.as_str());
+                    return Ok(Interpreter::StuckOnImport {
+                        state: self,
+                        module,
+                    });
+                }
+            }
+            self.document_stack[l].done_processing_imports();
+            self.document_stack[l].reorder(&self.bag)?;
+        }
+        let parsed_document = &mut self.document_stack[l];
 
         let mut instructions: Vec<ftd::Instruction> = Default::default();
-
-        for p1 in new_p1.iter() {
+        while let Some(p1) = parsed_document.sections.pop() {
             if p1.is_commented {
                 continue;
             }
 
-            if p1.name == "import" {
-                let (library_name, alias) =
-                    ftd::p2::utils::parse_import(&p1.caption, name, p1.line_number)?;
-                aliases.insert(alias, library_name.clone());
-                let start = std::time::Instant::now();
-                let doc = ftd::p2::TDoc {
-                    name,
-                    aliases: &aliases,
-                    bag: &self.bag,
-                    local_variables: &mut Default::default(),
-                };
-                let s = self
-                    .lib
-                    .get_with_result(library_name.as_str(), &doc)
-                    .await?;
-                *d_get = d_get.saturating_add(std::time::Instant::now() - start);
-                if !self.library_in_the_bag(library_name.as_str()) {
-                    self.async_interpret_(
-                        library_name.as_str(),
-                        s.as_str(),
-                        false,
-                        d_get,
-                        d_processor,
-                    )
-                    .await?;
-                    self.add_library_to_bag(library_name.as_str())
-                }
-                continue;
-            }
-
             // while this is a specific to entire document, we are still creating it in a loop
-            // because otherwise the self.interpret() call wont compile.
+            // because otherwise the self.interpret() call won't compile.
+
             let doc = ftd::p2::TDoc {
-                name,
-                aliases: &aliases,
+                name: &parsed_document.name,
+                aliases: &parsed_document.doc_aliases,
                 bag: &self.bag,
                 local_variables: &mut Default::default(),
             };
@@ -161,7 +80,7 @@ impl<'a> Interpreter<'a> {
                 &p1.name,
                 &doc,
                 p1.line_number,
-                &var_types,
+                &parsed_document.var_types,
             );
 
             let mut thing = vec![];
@@ -181,7 +100,7 @@ impl<'a> Interpreter<'a> {
                 thing.push((name, ftd::p2::Thing::Record(d)));
             } else if p1.name.starts_with("or-type ") {
                 // declare a record
-                let d = ftd::OrType::from_p1(p1, &doc)?;
+                let d = ftd::OrType::from_p1(&p1, &doc)?;
                 let name = doc.resolve_name(p1.line_number, &d.name.to_string())?;
                 if self.bag.contains_key(name.as_str()) {
                     return ftd::e2(
@@ -192,7 +111,7 @@ impl<'a> Interpreter<'a> {
                 }
                 thing.push((name, ftd::p2::Thing::OrType(d)));
             } else if p1.name.starts_with("map ") {
-                let d = ftd::Variable::map_from_p1(p1, &doc)?;
+                let d = ftd::Variable::map_from_p1(&p1, &doc)?;
                 let name = doc.resolve_name(p1.line_number, &d.name.to_string())?;
                 if self.bag.contains_key(name.as_str()) {
                     return ftd::e2(
@@ -220,7 +139,7 @@ impl<'a> Interpreter<'a> {
             }) = var_data
             {
                 // declare a function
-                let d = ftd::Component::from_p1(p1, &doc)?;
+                let d = ftd::Component::from_p1(&p1, &doc)?;
                 let name = doc.resolve_name(p1.line_number, &d.full_name.to_string())?;
                 if self.bag.contains_key(name.as_str()) {
                     return ftd::e2(
@@ -237,26 +156,17 @@ impl<'a> Interpreter<'a> {
                     .str(doc.name, p1.line_number, "$processor$")
                     .is_ok()
                 {
-                    let name = doc.resolve_name(p1.line_number, &var_data.name)?;
-                    let start = std::time::Instant::now();
-                    let value = self.lib.process(p1, &doc).await?;
-                    *d_processor = d_processor.saturating_add(std::time::Instant::now() - start);
-                    ftd::Variable {
-                        name,
-                        value: ftd::PropertyValue::Value { value },
-                        conditions: vec![],
-                        flags: ftd::variable::VariableFlags::from_p1(
-                            &p1.header,
-                            doc.name,
-                            p1.line_number,
-                        )?,
-                    }
+                    // processor case: 1
+                    return Ok(Interpreter::StuckOnProcessor {
+                        state: self,
+                        section: p1,
+                    });
                 } else if var_data.is_none() || var_data.is_optional() {
                     // declare and instantiate a variable
-                    ftd::Variable::from_p1(p1, &doc)?
+                    ftd::Variable::from_p1(&p1, &doc)?
                 } else {
                     // declare and instantiate a list
-                    ftd::Variable::list_from_p1(p1, &doc)?
+                    ftd::Variable::list_from_p1(&p1, &doc)?
                 };
                 let name = doc.resolve_name(p1.line_number, &d.name.to_string())?;
                 if self.bag.contains_key(name.as_str()) {
@@ -295,7 +205,7 @@ impl<'a> Interpreter<'a> {
                     );
                 }
                 if let Some(expr) = p1.header.str_optional(doc.name, p1.line_number, "if")? {
-                    let val = v.get_value(p1, &doc)?;
+                    let val = v.get_value(&p1, &doc)?;
                     v.conditions.push((
                         ftd::p2::Boolean::from_expression(
                             expr,
@@ -311,12 +221,17 @@ impl<'a> Interpreter<'a> {
                     .str_optional(doc.name, p1.line_number, "$processor$")?
                     .is_some()
                 {
-                    let start = std::time::Instant::now();
-                    let value = self.lib.process(p1, &doc).await?;
-                    *d_processor = d_processor.saturating_add(std::time::Instant::now() - start);
-                    v.value = ftd::PropertyValue::Value { value };
+                    // processor case: 2
+                    return Ok(Interpreter::StuckOnProcessor {
+                        state: self,
+                        section: p1.to_owned(),
+                    });
+                    // let start = std::time::Instant::now();
+                    // let value = self.lib.process(p1, &doc)?;
+                    // *d_processor = d_processor.saturating_add(std::time::Instant::now() - start);
+                    // v.value = ftd::PropertyValue::Value { value };
                 } else {
-                    v.update_from_p1(p1, &doc)?;
+                    v.update_from_p1(&p1, &doc)?;
                 }
                 thing.push((
                     doc.resolve_name(p1.line_number, doc_name.as_str())?,
@@ -333,18 +248,17 @@ impl<'a> Interpreter<'a> {
                         );
                     }
                     ftd::p2::Thing::Component(_) => {
-                        let p1 = {
-                            let mut p1 = p1.clone();
-                            if p1
-                                .header
-                                .str_optional(doc.name, p1.line_number, "$processor$")?
-                                .is_some()
-                            {
-                                let value = self.lib.process(&p1, &doc).await?;
-                                Self::p1_from_processor(&mut p1, value);
-                            }
-                            p1
-                        };
+                        if p1
+                            .header
+                            .str_optional(doc.name, p1.line_number, "$processor$")?
+                            .is_some()
+                        {
+                            // processor case: 3
+                            return Ok(Interpreter::StuckOnProcessor {
+                                state: self,
+                                section: p1.to_owned(),
+                            });
+                        }
                         if let Ok(loop_data) = p1.header.str(doc.name, p1.line_number, "$loop$") {
                             let section_to_subsection = ftd::p1::SubSection {
                                 name: p1.name.to_string(),
@@ -421,7 +335,7 @@ impl<'a> Interpreter<'a> {
                         }
                     }
                     ftd::p2::Thing::Record(mut r) => {
-                        r.add_instance(p1, &doc)?;
+                        r.add_instance(&p1, &doc)?;
                         thing.push((
                             doc.resolve_name(p1.line_number, &p1.name.to_string())?,
                             ftd::p2::Thing::Record(r),
@@ -448,409 +362,134 @@ impl<'a> Interpreter<'a> {
             self.bag.extend(thing);
         }
 
-        if is_main {
-            self.p1 = p1;
-            self.aliases = aliases;
+        if self.document_stack.len() > 1 {
+            return self.continue_after_pop();
         }
-        Ok(instructions)
+
+        Ok(Interpreter::Done {
+            state: self,
+            instructions,
+        })
     }
 
-    // #[cfg(not(feature = "async"))]
-    fn interpret_(
-        &mut self,
-        name: &str,
-        s: &str,
-        is_main: bool,
-        d_get: &mut std::time::Duration,
-        d_processor: &mut std::time::Duration,
-    ) -> ftd::p1::Result<Vec<ftd::Instruction>> {
-        let p1 = ftd::p1::parse(s, name)?;
-
-        // do all imports and then reorder
-        let mut aliases = default_aliases();
+    fn process_imports(
+        top: &mut ParsedDocument,
+        bag: &std::collections::BTreeMap<String, ftd::p2::Thing>,
+    ) -> ftd::p1::Result<Option<String>> {
         let mut iteration_index = 0;
-        while iteration_index < p1.len() && p1[iteration_index].name == "import" {
-            if p1[iteration_index].is_commented {
+        while iteration_index < top.sections.len() {
+            if top.sections[iteration_index].is_commented
+                || top.sections[iteration_index].name != "import"
+            {
                 iteration_index += 1;
                 continue;
             }
             let (library_name, alias) = ftd::p2::utils::parse_import(
-                &p1[iteration_index].caption,
-                name,
-                p1[iteration_index].line_number,
+                &top.sections[iteration_index].caption,
+                top.name.as_str(),
+                top.sections[iteration_index].line_number,
             )?;
-            aliases.insert(alias, library_name.clone());
-            let start = std::time::Instant::now();
-            let doc = ftd::p2::TDoc {
-                name,
-                aliases: &aliases,
-                bag: &self.bag,
-                local_variables: &mut Default::default(),
-            };
-            let s = self.lib.get_with_result(library_name.as_str(), &doc)?;
-            *d_get = d_get.saturating_add(std::time::Instant::now() - start);
-            if !self.library_in_the_bag(library_name.as_str()) {
-                self.interpret_(library_name.as_str(), s.as_str(), false, d_get, d_processor)?;
-                self.add_library_to_bag(library_name.as_str())
-            }
-            iteration_index += 1;
-        }
 
-        let (new_p1, var_types) = ftd::p2::utils::reorder(
-            &p1[iteration_index..],
-            &ftd::p2::TDoc {
-                name,
-                aliases: &aliases,
-                bag: &self.bag,
-                local_variables: &mut Default::default(),
-            },
-        )?;
+            top.doc_aliases.insert(alias, library_name.clone());
 
-        let mut instructions: Vec<ftd::Instruction> = Default::default();
-
-        for p1 in new_p1.iter() {
-            if p1.is_commented {
+            if bag.contains_key(library_name.as_str()) {
+                iteration_index += 1;
                 continue;
             }
 
-            if p1.name == "import" {
-                let (library_name, alias) =
-                    ftd::p2::utils::parse_import(&p1.caption, name, p1.line_number)?;
-                aliases.insert(alias, library_name.clone());
-                let start = std::time::Instant::now();
-                let doc = ftd::p2::TDoc {
-                    name,
-                    aliases: &aliases,
-                    bag: &self.bag,
-                    local_variables: &mut Default::default(),
-                };
-                let s = self.lib.get_with_result(library_name.as_str(), &doc)?;
-                *d_get = d_get.saturating_add(std::time::Instant::now() - start);
-                if !self.library_in_the_bag(library_name.as_str()) {
-                    self.interpret_(library_name.as_str(), s.as_str(), false, d_get, d_processor)?;
-                    self.add_library_to_bag(library_name.as_str())
-                }
-                continue;
-            }
-
-            // while this is a specific to entire document, we are still creating it in a loop
-            // because otherwise the self.interpret() call wont compile.
-            let doc = ftd::p2::TDoc {
-                name,
-                aliases: &aliases,
-                bag: &self.bag,
-                local_variables: &mut Default::default(),
-            };
-
-            let var_data = ftd::variable::VariableData::get_name_kind(
-                &p1.name,
-                &doc,
-                p1.line_number,
-                &var_types,
-            );
-
-            let mut thing = vec![];
-
-            if p1.name.starts_with("record ") {
-                // declare a record
-                let d =
-                    ftd::p2::Record::from_p1(p1.name.as_str(), &p1.header, &doc, p1.line_number)?;
-                let name = doc.resolve_name(p1.line_number, &d.name.to_string())?;
-                if self.bag.contains_key(name.as_str()) {
-                    return ftd::e2(
-                        format!("{} is already declared", d.name),
-                        doc.name,
-                        p1.line_number,
-                    );
-                }
-                thing.push((name, ftd::p2::Thing::Record(d)));
-            } else if p1.name.starts_with("or-type ") {
-                // declare a record
-                let d = ftd::OrType::from_p1(p1, &doc)?;
-                let name = doc.resolve_name(p1.line_number, &d.name.to_string())?;
-                if self.bag.contains_key(name.as_str()) {
-                    return ftd::e2(
-                        format!("{} is already declared", d.name),
-                        doc.name,
-                        p1.line_number,
-                    );
-                }
-                thing.push((name, ftd::p2::Thing::OrType(d)));
-            } else if p1.name.starts_with("map ") {
-                let d = ftd::Variable::map_from_p1(p1, &doc)?;
-                let name = doc.resolve_name(p1.line_number, &d.name.to_string())?;
-                if self.bag.contains_key(name.as_str()) {
-                    return ftd::e2(
-                        format!("{} is already declared", d.name),
-                        doc.name,
-                        p1.line_number,
-                    );
-                }
-                thing.push((name, ftd::p2::Thing::Variable(d)));
-                // } else if_two_words(p1.name.as_str() {
-                //   TODO: <record-name> <variable-name>: foo can be used to create a variable/
-                //         Not sure if its a good idea tho.
-                // }
-            } else if p1.name == "container" {
-                instructions.push(ftd::Instruction::ChangeContainer {
-                    name: doc.resolve_name_with_instruction(
-                        p1.line_number,
-                        p1.caption(p1.line_number, doc.name)?.as_str(),
-                        &instructions,
-                    )?,
-                });
-            } else if let Ok(ftd::variable::VariableData {
-                type_: ftd::variable::Type::Component,
-                ..
-            }) = var_data
-            {
-                // declare a function
-                let d = ftd::Component::from_p1(p1, &doc)?;
-                let name = doc.resolve_name(p1.line_number, &d.full_name.to_string())?;
-                if self.bag.contains_key(name.as_str()) {
-                    return ftd::e2(
-                        format!("{} is already declared", d.full_name),
-                        doc.name,
-                        p1.line_number,
-                    );
-                }
-                thing.push((name, ftd::p2::Thing::Component(d)));
-                // processed_p1.push(p1.name.to_string());
-            } else if let Ok(ref var_data) = var_data {
-                let d = if p1
-                    .header
-                    .str(doc.name, p1.line_number, "$processor$")
-                    .is_ok()
-                {
-                    let name = doc.resolve_name(p1.line_number, &var_data.name)?;
-                    let start = std::time::Instant::now();
-                    let value = self.lib.process(p1, &doc)?;
-                    *d_processor = d_processor.saturating_add(std::time::Instant::now() - start);
-                    ftd::Variable {
-                        name,
-                        value: ftd::PropertyValue::Value { value },
-                        conditions: vec![],
-                        flags: ftd::variable::VariableFlags::from_p1(
-                            &p1.header,
-                            doc.name,
-                            p1.line_number,
-                        )?,
-                    }
-                } else if var_data.is_none() || var_data.is_optional() {
-                    // declare and instantiate a variable
-                    ftd::Variable::from_p1(p1, &doc)?
-                } else {
-                    // declare and instantiate a list
-                    ftd::Variable::list_from_p1(p1, &doc)?
-                };
-                let name = doc.resolve_name(p1.line_number, &d.name.to_string())?;
-                if self.bag.contains_key(name.as_str()) {
-                    return ftd::e2(
-                        format!("{} is already declared", d.name),
-                        doc.name,
-                        p1.line_number,
-                    );
-                }
-                thing.push((name, ftd::p2::Thing::Variable(d)));
-            } else if let ftd::p2::Thing::Variable(mut v) =
-                doc.get_thing(p1.line_number, p1.name.as_str())?
-            {
-                assert!(
-                    !(p1.header
-                        .str_optional(doc.name, p1.line_number, "if")?
-                        .is_some()
-                        && p1
-                            .header
-                            .str_optional(doc.name, p1.line_number, "$processor$")?
-                            .is_some())
-                );
-                let (doc_name, remaining) = ftd::p2::utils::get_doc_name_and_remaining(
-                    doc.resolve_name(p1.line_number, p1.name.as_str())?.as_str(),
-                )?;
-                if remaining.is_some()
-                    && p1
-                        .header
-                        .str_optional(doc.name, p1.line_number, "if")?
-                        .is_some()
-                {
-                    return ftd::e2(
-                        "Currently not supporting `if` for field value update.",
-                        doc.name,
-                        p1.line_number,
-                    );
-                }
-                if let Some(expr) = p1.header.str_optional(doc.name, p1.line_number, "if")? {
-                    let val = v.get_value(p1, &doc)?;
-                    v.conditions.push((
-                        ftd::p2::Boolean::from_expression(
-                            expr,
-                            &doc,
-                            &Default::default(),
-                            (None, None),
-                            p1.line_number,
-                        )?,
-                        val,
-                    ));
-                } else if p1
-                    .header
-                    .str_optional(doc.name, p1.line_number, "$processor$")?
-                    .is_some()
-                {
-                    let start = std::time::Instant::now();
-                    let value = self.lib.process(p1, &doc)?;
-                    *d_processor = d_processor.saturating_add(std::time::Instant::now() - start);
-                    v.value = ftd::PropertyValue::Value { value };
-                } else {
-                    v.update_from_p1(p1, &doc)?;
-                }
-                thing.push((
-                    doc.resolve_name(p1.line_number, doc_name.as_str())?,
-                    ftd::p2::Thing::Variable(doc.set_value(p1.line_number, p1.name.as_str(), v)?),
-                ));
-            } else {
-                // cloning because https://github.com/rust-lang/rust/issues/59159
-                match (doc.get_thing(p1.line_number, p1.name.as_str())?).clone() {
-                    ftd::p2::Thing::Variable(_) => {
-                        return ftd::e2(
-                            format!("variable should have prefix $, found: `{}`", p1.name),
-                            doc.name,
-                            p1.line_number,
-                        );
-                    }
-                    ftd::p2::Thing::Component(_) => {
-                        let p1 = {
-                            let mut p1 = p1.clone();
-                            if p1
-                                .header
-                                .str_optional(doc.name, p1.line_number, "$processor$")?
-                                .is_some()
-                            {
-                                let value = self.lib.process(&p1, &doc)?;
-                                Self::p1_from_processor(&mut p1, value);
-                            }
-                            p1
-                        };
-                        if let Ok(loop_data) = p1.header.str(doc.name, p1.line_number, "$loop$") {
-                            let section_to_subsection = ftd::p1::SubSection {
-                                name: p1.name.to_string(),
-                                caption: p1.caption.to_owned(),
-                                header: p1.header.to_owned(),
-                                body: p1.body.to_owned(),
-                                is_commented: p1.is_commented,
-                                line_number: p1.line_number,
-                            };
-                            instructions.push(ftd::Instruction::RecursiveChildComponent {
-                                child: ftd::component::recursive_child_component(
-                                    loop_data,
-                                    &section_to_subsection,
-                                    &doc,
-                                    &Default::default(),
-                                    None,
-                                )?,
-                            });
-                        } else {
-                            let parent = ftd::ChildComponent::from_p1(
-                                p1.line_number,
-                                p1.name.as_str(),
-                                &p1.header,
-                                &p1.caption,
-                                &p1.body_without_comment(),
-                                &doc,
-                                &Default::default(),
-                            )?;
-
-                            let mut children = vec![];
-
-                            for sub in p1.sub_sections.0.iter() {
-                                if sub.is_commented {
-                                    continue;
-                                }
-                                if let Ok(loop_data) =
-                                    sub.header.str(doc.name, p1.line_number, "$loop$")
-                                {
-                                    children.push(ftd::component::recursive_child_component(
-                                        loop_data,
-                                        sub,
-                                        &doc,
-                                        &parent.arguments,
-                                        None,
-                                    )?);
-                                } else {
-                                    let root_name = ftd::p2::utils::get_root_component_name(
-                                        &doc,
-                                        parent.root.as_str(),
-                                        sub.line_number,
-                                    )?;
-                                    let child = if root_name.eq("ftd#text") {
-                                        ftd::p2::utils::get_markup_child(
-                                            sub,
-                                            &doc,
-                                            &parent.arguments,
-                                        )?
-                                    } else {
-                                        ftd::ChildComponent::from_p1(
-                                            sub.line_number,
-                                            sub.name.as_str(),
-                                            &sub.header,
-                                            &sub.caption,
-                                            &sub.body_without_comment(),
-                                            &doc,
-                                            &parent.arguments,
-                                        )?
-                                    };
-                                    children.push(child);
-                                }
-                            }
-
-                            instructions.push(ftd::Instruction::Component { children, parent })
-                        }
-                    }
-                    ftd::p2::Thing::Record(mut r) => {
-                        r.add_instance(p1, &doc)?;
-                        thing.push((
-                            doc.resolve_name(p1.line_number, &p1.name.to_string())?,
-                            ftd::p2::Thing::Record(r),
-                        ));
-                    }
-                    ftd::p2::Thing::OrType(_r) => {
-                        // do we allow initialization of a record by name? nopes
-                        return ftd::e2(
-                            format!("'{}' is an or-type", p1.name.as_str()),
-                            doc.name,
-                            p1.line_number,
-                        );
-                    }
-                    ftd::p2::Thing::OrTypeWithVariant { .. } => {
-                        // do we allow initialization of a record by name? nopes
-                        return ftd::e2(
-                            format!("'{}' is an or-type variant", p1.name.as_str(),),
-                            doc.name,
-                            p1.line_number,
-                        );
-                    }
-                };
-            }
-            self.bag.extend(thing);
+            top.sections.remove(iteration_index);
+            return Ok(Some(library_name));
         }
 
-        if is_main {
-            self.p1 = p1;
-            self.aliases = aliases;
-        }
-        Ok(instructions)
+        Ok(None)
     }
 
-    pub(crate) fn new(lib: &'a dyn ftd::p2::Library) -> Self {
-        Self {
-            lib,
-            bag: default_bag(),
-            p1: Default::default(),
-            aliases: Default::default(),
-            parsed_libs: Default::default(),
+    pub fn continue_after_import(mut self, id: &str, source: &str) -> ftd::p1::Result<Interpreter> {
+        self.document_stack.push(ParsedDocument::parse(id, source)?);
+        self.continue_()
+        // interpret then
+        // handle top / start_from
+    }
+
+    pub fn continue_after_pop(mut self) -> ftd::p1::Result<Interpreter> {
+        self.document_stack.pop();
+        self.continue_()
+        // interpret then
+        // handle top / start_from
+    }
+
+    pub fn continue_after_processor(
+        mut self,
+        p1: &ftd::p1::Section,
+        lib: &dyn ftd::p2::Library,
+    ) -> ftd::p1::Result<Interpreter> {
+        let l = self.document_stack.len() - 1;
+        let parsed_document = &mut self.document_stack[l];
+        let doc = ftd::p2::TDoc {
+            name: &parsed_document.name,
+            aliases: &parsed_document.doc_aliases,
+            bag: &self.bag,
+            local_variables: &mut Default::default(),
+        };
+        let value = lib.process(p1, &doc)?;
+
+        let var_data = ftd::variable::VariableData::get_name_kind(
+            &p1.name,
+            &doc,
+            p1.line_number,
+            &parsed_document.var_types,
+        );
+
+        if let Ok(ftd::variable::VariableData {
+            type_: ftd::variable::Type::Variable,
+            name,
+            ..
+        }) = var_data
+        {
+            let name = doc.resolve_name(p1.line_number, &name)?;
+            let variable = ftd::p2::Thing::Variable(ftd::Variable {
+                name: name.clone(),
+                value: ftd::PropertyValue::Value { value },
+                conditions: vec![],
+                flags: ftd::variable::VariableFlags::from_p1(&p1.header, doc.name, p1.line_number)?,
+            });
+            self.bag.insert(name, variable);
+            return self.continue_();
         }
+
+        match doc.get_thing(p1.line_number, p1.name.as_str())? {
+            ftd::p2::Thing::Variable(mut v) => {
+                // for case: 2
+                let doc_name = ftd::p2::utils::get_doc_name_and_remaining(
+                    doc.resolve_name(p1.line_number, p1.name.as_str())?.as_str(),
+                )?
+                .0;
+                v.value = ftd::PropertyValue::Value { value };
+                let key = doc.resolve_name(p1.line_number, doc_name.as_str())?;
+                let variable =
+                    ftd::p2::Thing::Variable(doc.set_value(p1.line_number, p1.name.as_str(), v)?);
+                self.bag.insert(key, variable);
+            }
+            ftd::p2::Thing::Component(_) => {
+                // for case: 3
+                let mut p1 = p1.clone();
+                Self::p1_from_processor(&mut p1, value);
+                parsed_document.sections.push(p1.to_owned());
+            }
+            _ => todo!(), // throw error
+        }
+        self.continue_()
+        // interpret then
+        // handle top / start_from
     }
 
     pub(crate) fn p1_from_processor(p1: &mut ftd::p1::Section, value: ftd::Value) {
+        for (idx, (_, k, _)) in p1.header.0.iter().enumerate() {
+            if k.eq("$processor$") {
+                p1.header.0.remove(idx);
+                break;
+            }
+        }
         if let ftd::Value::Object { values } = value {
             for (k, v) in values {
                 let v = if let ftd::PropertyValue::Value { value } = v {
@@ -875,17 +514,108 @@ impl<'a> Interpreter<'a> {
     }
 }
 
-pub fn interpret(
+#[derive(Debug, Clone)]
+pub struct ParsedDocument {
+    name: String,
+    sections: Vec<ftd::p1::Section>,
+    processing_imports: bool,
+    doc_aliases: std::collections::BTreeMap<String, String>,
+    var_types: Vec<String>,
+}
+
+impl ParsedDocument {
+    fn parse(id: &str, source: &str) -> ftd::p1::Result<ParsedDocument> {
+        Ok(ParsedDocument {
+            name: id.to_string(),
+            sections: ftd::p1::parse(source, id)?,
+            processing_imports: true,
+            doc_aliases: ftd::p2::interpreter::default_aliases(),
+            var_types: Default::default(),
+        })
+    }
+
+    fn done_processing_imports(&mut self) {
+        self.processing_imports = false;
+    }
+
+    fn reorder(
+        &mut self,
+        bag: &std::collections::BTreeMap<String, ftd::p2::Thing>,
+    ) -> ftd::p1::Result<()> {
+        let (mut new_p1, var_types) = ftd::p2::utils::reorder(
+            &self.sections,
+            &ftd::p2::TDoc {
+                name: &self.name,
+                aliases: &self.doc_aliases,
+                bag,
+                local_variables: &mut Default::default(),
+            },
+        )?;
+        new_p1.reverse();
+        self.sections = new_p1;
+        self.var_types = var_types;
+        Ok(())
+    }
+}
+
+#[derive(Debug)]
+pub enum Interpreter {
+    StuckOnImport {
+        module: String,
+        state: InterpreterState,
+    },
+    StuckOnProcessor {
+        state: InterpreterState,
+        section: ftd::p1::Section,
+    },
+    Done {
+        state: InterpreterState,
+        instructions: Vec<ftd::Instruction>,
+    },
+}
+
+pub fn interpret(id: &str, source: &str) -> ftd::p1::Result<Interpreter> {
+    let mut s = InterpreterState::new();
+    s.document_stack.push(ParsedDocument::parse(id, source)?);
+    s.continue_()
+}
+
+pub fn interpret_helper(
     name: &str,
     source: &str,
     lib: &dyn ftd::p2::Library,
-) -> ftd::p1::Result<(
-    std::collections::BTreeMap<String, ftd::p2::Thing>,
-    ftd::Column,
-)> {
-    let mut rt = ftd::p2::interpreter2::interpret_helper(name, source, lib)?;
-    let main = rt.render_()?;
-    Ok((rt.bag, main))
+) -> ftd::p1::Result<ftd::RT> {
+    let mut s = interpret(name, source)?;
+    let instructions: Vec<ftd::Instruction>;
+    let state;
+    loop {
+        match s {
+            Interpreter::Done {
+                instructions: i,
+                state: st,
+            } => {
+                instructions = i;
+                state = st;
+                break;
+            }
+            Interpreter::StuckOnProcessor { state, section } => {
+                s = state.continue_after_processor(&section, lib)?;
+            }
+            Interpreter::StuckOnImport { module, state: st } => {
+                let mut bt: std::collections::BTreeMap<String, ftd::p2::Thing> =
+                    std::collections::BTreeMap::new();
+                let source = lib.get_with_result(module.as_str(), &st.tdoc(&mut bt))?;
+                s = st.continue_after_import(module.as_str(), source.as_str())?;
+            }
+        }
+    }
+
+    Ok(ftd::RT::from(
+        name,
+        state.document_stack[0].clone().doc_aliases,
+        state.bag,
+        instructions,
+    ))
 }
 
 #[allow(clippy::large_enum_variant)]
@@ -4138,7 +3868,7 @@ mod test {
                 ..Default::default()
             }));
 
-        let (g_bag, g_col) = ftd::p2::interpreter::interpret(
+        let (g_bag, g_col) = ftd::test::interpret(
             "foo/bar",
             indoc::indoc!(
                 "
@@ -7739,7 +7469,7 @@ mod test {
                 ..Default::default()
             }));
 
-        let (_g_bag, g_col) = ftd::p2::interpreter::interpret(
+        let (_g_bag, g_col) = ftd::test::interpret(
             "foo/bar",
             indoc::indoc!(
                 "
@@ -7872,7 +7602,7 @@ mod test {
                 ..Default::default()
             }));
 
-        let (_g_bag, g_col) = ftd::p2::interpreter::interpret(
+        let (_g_bag, g_col) = ftd::test::interpret(
             "foo/bar",
             indoc::indoc!(
                 "
@@ -8068,7 +7798,7 @@ mod test {
                 ..Default::default()
             }));
 
-        let (_g_bag, g_col) = ftd::p2::interpreter::interpret(
+        let (_g_bag, g_col) = ftd::test::interpret(
             "foo/bar",
             indoc::indoc!(
                 "
@@ -8237,7 +7967,7 @@ mod test {
                 ..Default::default()
             }));
 
-        let (_g_bag, g_col) = ftd::p2::interpreter::interpret(
+        let (_g_bag, g_col) = ftd::test::interpret(
             "foo/bar",
             indoc::indoc!(
                 "
@@ -8333,7 +8063,7 @@ mod test {
                 ..Default::default()
             },
         }));
-        let (g_bag, g_col) = ftd::p2::interpreter::interpret(
+        let (g_bag, g_col) = ftd::test::interpret(
             "foo/bar",
             indoc::indoc!(
                 "
@@ -8375,7 +8105,7 @@ mod test {
                 ..Default::default()
             }));
 
-        let (g_bag, g_col) = ftd::p2::interpreter::interpret(
+        let (g_bag, g_col) = ftd::test::interpret(
             "foo/bar",
             indoc::indoc!(
                 "
@@ -9385,7 +9115,7 @@ mod test {
                 conditions: vec![],
             }),
         );
-        let (g_bag, g_col) = ftd::p2::interpreter::interpret(
+        let (g_bag, g_col) = ftd::test::interpret(
             "foo/bar",
             indoc::indoc!(
                 "
@@ -9661,7 +9391,7 @@ mod test {
             }),
         );
 
-        let (_g_bag, g_col) = ftd::p2::interpreter::interpret(
+        let (_g_bag, g_col) = ftd::test::interpret(
             "foo/bar",
             indoc::indoc!(
                 "
@@ -9734,7 +9464,7 @@ mod test {
             }),
         );
 
-        let (g_bag, g_col) = ftd::p2::interpreter::interpret(
+        let (g_bag, g_col) = ftd::test::interpret(
             "foo/bar",
             indoc::indoc!(
                 "
@@ -9784,7 +9514,7 @@ mod test {
             }),
         );
 
-        let (g_bag, g_col) = ftd::p2::interpreter::interpret(
+        let (g_bag, g_col) = ftd::test::interpret(
             "foo/bar",
             indoc::indoc!(
                 "
@@ -10112,7 +9842,7 @@ mod test {
             }),
         );
 
-        let (g_bag, g_col) = ftd::p2::interpreter::interpret(
+        let (g_bag, g_col) = ftd::test::interpret(
             "foo/bar",
             indoc::indoc!(
                 "
@@ -10686,7 +10416,7 @@ mod test {
             }),
         );*/
 
-        let (_g_bag, g_col) = ftd::p2::interpreter::interpret(
+        let (_g_bag, g_col) = ftd::test::interpret(
             "foo/bar",
             indoc::indoc!(
                 "
@@ -11165,7 +10895,7 @@ mod test {
             }),
         );
 
-        let (_g_bag, g_col) = ftd::p2::interpreter::interpret(
+        let (_g_bag, g_col) = ftd::test::interpret(
             "foo/bar",
             indoc::indoc!(
                 "
@@ -11502,7 +11232,7 @@ mod test {
                 flags: Default::default(),
             }),
         );
-        let (g_bag, g_col) = ftd::p2::interpreter::interpret(
+        let (g_bag, g_col) = ftd::test::interpret(
             "foo/bar",
             indoc::indoc!(
                 "
@@ -11659,7 +11389,7 @@ mod test {
                 ..Default::default()
             }));
 
-        let (g_bag, g_col) = ftd::p2::interpreter::interpret(
+        let (g_bag, g_col) = ftd::test::interpret(
             "foo/bar",
             indoc::indoc!(
                 "
@@ -12065,7 +11795,7 @@ mod test {
             }),
         );
 
-        let (g_bag, g_col) = ftd::p2::interpreter::interpret(
+        let (g_bag, g_col) = ftd::test::interpret(
             "foo/bar",
             indoc::indoc!(
                 "
@@ -12220,7 +11950,7 @@ mod test {
             ..Default::default()
         }));
 
-        let (_g_bag, g_col) = ftd::p2::interpreter::interpret(
+        let (_g_bag, g_col) = ftd::test::interpret(
             "foo/bar",
             indoc::indoc!(
                 "
@@ -12479,7 +12209,7 @@ mod test {
                 },
             }));
 
-        let (_g_bag, g_col) = ftd::p2::interpreter::interpret(
+        let (_g_bag, g_col) = ftd::test::interpret(
             "foo/bar",
             indoc::indoc!(
                 "
@@ -12598,7 +12328,7 @@ mod test {
                 ..Default::default()
             }));
 
-        let (_g_bag, g_col) = ftd::p2::interpreter::interpret(
+        let (_g_bag, g_col) = ftd::test::interpret(
             "foo/bar",
             indoc::indoc!(
                 "
@@ -12744,7 +12474,7 @@ mod test {
             }),
         );
 
-        let (g_bag, g_col) = ftd::p2::interpreter::interpret(
+        let (g_bag, g_col) = ftd::test::interpret(
             "foo/bar",
             indoc::indoc!(
                 "
@@ -12822,7 +12552,7 @@ mod test {
                 ..Default::default()
             }));
 
-        let (_g_bag, g_col) = ftd::p2::interpreter::interpret(
+        let (_g_bag, g_col) = ftd::test::interpret(
             "foo/bar",
             indoc::indoc!(
                 "
@@ -12943,7 +12673,7 @@ mod test {
                 },
             }));
 
-        let (_g_bag, g_col) = ftd::p2::interpreter::interpret(
+        let (_g_bag, g_col) = ftd::test::interpret(
             "foo/bar",
             indoc::indoc!(
                 "
@@ -13062,7 +12792,7 @@ mod test {
                 ..Default::default()
             }));
 
-        let (_g_bag, g_col) = ftd::p2::interpreter::interpret(
+        let (_g_bag, g_col) = ftd::test::interpret(
             "foo/bar",
             indoc::indoc!(
                 "
@@ -13114,7 +12844,7 @@ mod test {
                 ..Default::default()
             }));
 
-        let (_g_bag, g_col) = ftd::p2::interpreter::interpret(
+        let (_g_bag, g_col) = ftd::test::interpret(
             "foo/bar",
             indoc::indoc!(
                 "
@@ -13152,7 +12882,7 @@ mod test {
                 ..Default::default()
             }));
 
-        let (_g_bag, g_col) = ftd::p2::interpreter::interpret(
+        let (_g_bag, g_col) = ftd::test::interpret(
             "foo/bar",
             indoc::indoc!(
                 "
@@ -13194,7 +12924,7 @@ mod test {
                 ..Default::default()
             }));
 
-        let (_g_bag, g_col) = ftd::p2::interpreter::interpret(
+        let (_g_bag, g_col) = ftd::test::interpret(
             "foo/bar",
             indoc::indoc!(
                 "
@@ -13291,7 +13021,7 @@ mod test {
             ..Default::default()
         }));
 
-        let (_g_bag, g_col) = ftd::p2::interpreter::interpret(
+        let (_g_bag, g_col) = ftd::test::interpret(
             "foo/bar",
             indoc::indoc!(
                 r"
@@ -13401,7 +13131,7 @@ mod test {
                 ..Default::default()
             }));
 
-        let (_g_bag, g_col) = ftd::p2::interpreter::interpret(
+        let (_g_bag, g_col) = ftd::test::interpret(
             "foo/bar",
             indoc::indoc!(
                 "
@@ -13599,7 +13329,7 @@ mod test {
                 ..Default::default()
             }));
 
-        let (_g_bag, g_col) = ftd::p2::interpreter::interpret(
+        let (_g_bag, g_col) = ftd::test::interpret(
             "foo/bar",
             indoc::indoc!(
                 "
@@ -13702,7 +13432,7 @@ mod test {
                 ..Default::default()
             }));
 
-        let (_g_bag, g_col) = ftd::p2::interpreter::interpret(
+        let (_g_bag, g_col) = ftd::test::interpret(
             "foo/bar",
             indoc::indoc!(
                 "
@@ -13752,7 +13482,7 @@ mod test {
             ..Default::default()
         }));
 
-        let (_g_bag, g_col) = ftd::p2::interpreter::interpret(
+        let (_g_bag, g_col) = ftd::test::interpret(
             "foo/bar",
             indoc::indoc!(
                 "
@@ -13854,7 +13584,7 @@ mod test {
                 ..Default::default()
             }));
 
-        let (_g_bag, g_col) = ftd::p2::interpreter::interpret(
+        let (_g_bag, g_col) = ftd::test::interpret(
             "foo/bar",
             indoc::indoc!(
                 "
@@ -13933,7 +13663,7 @@ mod test {
                 ..Default::default()
             }));
 
-        let (_g_bag, g_col) = ftd::p2::interpreter::interpret(
+        let (_g_bag, g_col) = ftd::test::interpret(
             "foo/bar",
             indoc::indoc!(
                 "
@@ -14190,7 +13920,7 @@ mod test {
                 ..Default::default()
             }));
 
-        let (_g_bag, g_col) = ftd::p2::interpreter::interpret(
+        let (_g_bag, g_col) = ftd::test::interpret(
             "foo/bar",
             indoc::indoc!(
                 "
@@ -14279,7 +14009,7 @@ mod test {
                 ..Default::default()
             }));
 
-        let (_g_bag, g_col) = ftd::p2::interpreter::interpret(
+        let (_g_bag, g_col) = ftd::test::interpret(
             "foo/bar",
             indoc::indoc!(
                 r"
@@ -14332,7 +14062,7 @@ mod test {
                 ..Default::default()
             }));
 
-        let (_g_bag, g_col) = ftd::p2::interpreter::interpret(
+        let (_g_bag, g_col) = ftd::test::interpret(
             "foo/bar",
             indoc::indoc!(
                 "
@@ -14435,7 +14165,7 @@ mod test {
                 },
             }));
 
-        let (_g_bag, g_col) = ftd::p2::interpreter::interpret(
+        let (_g_bag, g_col) = ftd::test::interpret(
             "foo/bar",
             indoc::indoc!(
                 "
@@ -14520,7 +14250,7 @@ mod test {
                 },
             }));
 
-        let (_g_bag, g_col) = ftd::p2::interpreter::interpret(
+        let (_g_bag, g_col) = ftd::test::interpret(
             "foo/bar",
             indoc::indoc!(
                 "
@@ -14590,7 +14320,7 @@ mod test {
                 },
                 ..Default::default()
             }));
-        let (_g_bag, g_col) = ftd::p2::interpreter::interpret(
+        let (_g_bag, g_col) = ftd::test::interpret(
             "foo/bar",
             indoc::indoc!(
                 "
@@ -14654,7 +14384,7 @@ mod test {
                 },
             }));
 
-        let (_g_bag, g_col) = ftd::p2::interpreter::interpret(
+        let (_g_bag, g_col) = ftd::test::interpret(
             "foo/bar",
             indoc::indoc!(
                 "
@@ -14739,7 +14469,7 @@ mod test {
                 ..Default::default()
             }));
 
-        let (_g_bag, g_col) = ftd::p2::interpreter::interpret(
+        let (_g_bag, g_col) = ftd::test::interpret(
             "foo/bar",
             indoc::indoc!(
                 "
@@ -14847,7 +14577,7 @@ mod test {
                 ..Default::default()
             }));
 
-        let (_g_bag, g_col) = ftd::p2::interpreter::interpret(
+        let (_g_bag, g_col) = ftd::test::interpret(
             "foo/bar",
             indoc::indoc!(
                 "
@@ -14965,7 +14695,7 @@ mod test {
                 }
             }));
 
-        let (_g_bag, g_col) = ftd::p2::interpreter::interpret(
+        let (_g_bag, g_col) = ftd::test::interpret(
             "foo/bar",
             indoc::indoc!(
                 "
@@ -15092,7 +14822,7 @@ mod test {
                 ..Default::default()
             }));
 
-        let (_g_bag, g_col) = ftd::p2::interpreter::interpret(
+        let (_g_bag, g_col) = ftd::test::interpret(
             "foo/bar",
             indoc::indoc!(
                 "
@@ -15135,7 +14865,7 @@ mod test {
                 ..Default::default()
             }));
 
-        let (_g_bag, g_col) = ftd::p2::interpreter::interpret(
+        let (_g_bag, g_col) = ftd::test::interpret(
             "foo/bar",
             indoc::indoc!(
                 "
@@ -15171,7 +14901,7 @@ mod test {
                 ..Default::default()
             }));
 
-        let (_g_bag, g_col) = ftd::p2::interpreter::interpret(
+        let (_g_bag, g_col) = ftd::test::interpret(
             "foo/bar",
             indoc::indoc!(
                 "
@@ -15228,7 +14958,7 @@ mod test {
         main.container.children.push(col.clone());
         main.container.children.push(col);
 
-        let (_g_bag, g_col) = ftd::p2::interpreter::interpret(
+        let (_g_bag, g_col) = ftd::test::interpret(
             "foo/bar",
             indoc::indoc!(
                 "
@@ -15341,7 +15071,7 @@ mod test {
                 ..Default::default()
             }));
 
-        let (_g_bag, g_col) = ftd::p2::interpreter::interpret(
+        let (_g_bag, g_col) = ftd::test::interpret(
             "foo/bar",
             indoc::indoc!(
                 "
@@ -15437,7 +15167,7 @@ mod test {
                 },
             }));
 
-        let (_g_bag, g_col) = ftd::p2::interpreter::interpret(
+        let (_g_bag, g_col) = ftd::test::interpret(
             "foo/bar",
             indoc::indoc!(
                 "
@@ -15547,7 +15277,7 @@ mod test {
             },
         }));
 
-        let (_g_bag, g_col) = ftd::p2::interpreter::interpret(
+        let (_g_bag, g_col) = ftd::test::interpret(
             "foo/bar",
             indoc::indoc!(
                 "
@@ -15615,7 +15345,7 @@ mod test {
             ..Default::default()
         }));
 
-        let (_g_bag, g_col) = ftd::p2::interpreter::interpret(
+        let (_g_bag, g_col) = ftd::test::interpret(
             "foo/bar",
             indoc::indoc!(
                 "
@@ -15781,7 +15511,7 @@ mod test {
             },
         }));
 
-        let (_g_bag, g_col) = ftd::p2::interpreter::interpret(
+        let (_g_bag, g_col) = ftd::test::interpret(
             "foo/bar",
             indoc::indoc!(
                 "
@@ -15903,7 +15633,7 @@ mod test {
                 ..Default::default()
             }));
 
-        let (_g_bag, g_col) = ftd::p2::interpreter::interpret(
+        let (_g_bag, g_col) = ftd::test::interpret(
             "foo/bar",
             indoc::indoc!(
                 "
@@ -15970,7 +15700,7 @@ mod test {
             ..Default::default()
         }));
 
-        let (_g_bag, g_col) = ftd::p2::interpreter::interpret(
+        let (_g_bag, g_col) = ftd::test::interpret(
             "foo/bar",
             indoc::indoc!(
                 "
@@ -16112,7 +15842,7 @@ mod test {
                 ..Default::default()
             }));
 
-        let (_g_bag, g_col) = crate::p2::interpreter::interpret(
+        let (_g_bag, g_col) = ftd::test::interpret(
             "foo/bar",
             indoc::indoc!(
                 "
@@ -16229,7 +15959,7 @@ mod test {
                 ..Default::default()
             }));
 
-        let (_g_bag, g_col) = crate::p2::interpreter::interpret(
+        let (_g_bag, g_col) = ftd::test::interpret(
             "foo/bar",
             indoc::indoc!(
                 "
@@ -16919,7 +16649,7 @@ mod test {
             }),
         );
 
-        let (g_bag, g_col) = crate::p2::interpreter::interpret(
+        let (g_bag, g_col) = ftd::test::interpret(
             "foo/bar",
             indoc::indoc!(
                 "
@@ -17604,7 +17334,7 @@ mod test {
 
     /*#[test]
     fn loop_with_tree_structure_1() {
-        let (g_bag, g_col) = ftd::p2::interpreter::interpret(
+        let (g_bag, g_col) = ftd::test::interpret(
             "foo/bar",
             indoc::indoc!(
                 "
@@ -17667,7 +17397,7 @@ mod test {
 
     #[test]
     fn loop_with_tree_structure_2() {
-        let (g_bag, g_col) = ftd::p2::interpreter::interpret(
+        let (g_bag, g_col) = ftd::test::interpret(
             "foo/bar",
             indoc::indoc!(
                 "
