+++ src/p2/mod.rs
@@ -3,7 +3,6 @@ pub(crate) mod element;
 pub(crate) mod event;
 pub(crate) mod expression;
 pub(crate) mod interpreter;
-pub mod interpreter2;
 pub mod interpreter3;
 pub(crate) mod kind;
 pub(crate) mod library;
